indexing_pipeline/requirements.txt
fastapi==0.110.1
pyvi==0.1.1
requests==2.32.3
uvicorn==0.29.0
weaviate-client==3.26.6
python-multipart
numpy